Update README.md
This contains homework for Module 3 - Python.  PyBank reads a csv file that sums all the profits and losses in a period of time, including the greatest increase month over month, and greatest decrease.  PyPoll identifies the candidates voted for in an election and the number of votes they received.  It also identifies the winner.